+++ spring-addons-starter-rest/src/main/java/com/c4_soft/springaddons/rest/AbstractWebClientBuilderFactoryBean.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import javax.net.ssl.SSLException;
 import org.jspecify.annotations.Nullable;
 import org.springframework.beans.factory.FactoryBean;
+import org.springframework.boot.webclient.autoconfigure.WebClientSsl;
 import org.springframework.http.client.reactive.ReactorClientHttpConnector;
 import org.springframework.web.reactive.function.client.ClientRequest;
 import org.springframework.web.reactive.function.client.ExchangeFilterFunction;
@@ -36,6 +37,7 @@ public abstract class AbstractWebClientBuilderFactoryBean
   private SystemProxyProperties systemProxyProperties = new SystemProxyProperties();
   private SpringAddonsRestProperties restProperties = new SpringAddonsRestProperties();
   private WebClient.Builder webClientBuilder;
+  private Optional<WebClientSsl> ssl = Optional.empty();
 
 
   @Override
@@ -54,6 +56,8 @@ public abstract class AbstractWebClientBuilderFactoryBean
       builder.defaultHeader(header.getKey(),
           header.getValue().toArray(new String[header.getValue().size()]));
     }
+
+    clientProps.getSslBundle().ifPresent(bundleName -> ssl.get().fromBundle(bundleName));
 
     return builder;
   }